ProjectOCR/requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pdf2image==0.1.4
Pillow==6.1.0
pytesseract==0.3.0
Werkzeug==0.15.6
